+++ core/impl/src/main/java/com/blazebit/persistence/impl/CachingJpaProvider.java
@@ -418,6 +418,11 @@ public final class CachingJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean supportsProxyParameterForNonPkAssociation() {
+        return jpaProvider.supportsProxyParameterForNonPkAssociation();
+    }
+
+    @Override
     public Object getIdentifier(Object entity) {
         return jpaProvider.getIdentifier(entity);
     }